+++ src/components/Home.js
@@ -1,40 +1,13 @@
-import React, { Component } from 'react';
-import { Media } from 'reactstrap';
+import React from 'react';
+import { Container } from 'reactstrap';
 
-function RenderMedia({item}) {
+function Home(){
   return(
-    <div className="mb-4 pt-2 col-12 offset-md-1 col-md-9 col-lg-8 educationMedia" 
-      style={{backgroundColor: window.innerWidth < 995 ? 'white' : ''}}>
-      <Media className="pb-2">
-        <Media body>
-          <Media heading>{item.name}</Media>
-          <Media heading>{item.shortname}</Media>
-          <p>{item.branch}</p>
-          <p> Graduated in {item.year}</p>
-        </Media>
-        <div className="d-none d-sm-block" >
-        <a href={item.href} target="_blank" rel="noopener noreferrer">
-          <Media style={{borderRadius: '50px'}} object src={item.src} alt={item.name} width="100px" height="120px" />
-        </a>
-        </div>
-      </Media>
-    </div>
+    <Container className="text-center">
+      <img className="mb-4" src="/assets/images/onclouds.jpg" alt="Bhushan Mhatre" style={{borderRadius: '50%'}}/>
+      <h2><i>Bhushan Chandrakant Mhatre</i></h2>
+    </Container>
   );
-}
-
-class Home extends Component{
-
-  render(){
-    const institutes = this.props.education.map(item=>
-        <RenderMedia item={item} />
-    )
-    return(
-      <div className="container">
-        <h2><span className="fa fa-institution fa-md mb-4"> Education</span></h2>
-        {institutes}
-      </div>
-    );
-  }
 }
 
 export default Home;
