+++ src/components/Navigation.js
@@ -30,6 +30,7 @@ class Navigation extends Component {
               <NavItem><NavLink href='/workexp'><span className="fa fa-server fa-lg"> Work-Experience</span></NavLink></NavItem>
               <NavItem><NavLink href='/project'><span className="fa fa-code fa-lg"> Projects</span></NavLink></NavItem>
               <NavItem><NavLink href='/certificate'><span className="fa fa-trophy fa-lg"> Certificates</span></NavLink></NavItem>
+              <NavItem><NavLink href='/education'><span className="fa fa-institution fa-lg"> Education</span></NavLink></NavItem>
               <NavItem><NavLink href='/contact'><span className="fa fa-phone fa-lg"> Contact</span></NavLink></NavItem>
             </Nav>
           </Collapse>
